Django-BE-CareerMentor/requirements.txt
django-rest-knox
django-cors-headers
markdown
django-filter
djangorestframework
Django
joblib